% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear_GFisher_cache}
\alias{clear_GFisher_cache}
\title{Clear GFisher Coefficient Cache}
\usage{
clear_GFisher_cache()
}
\value{
The number of entries that were cleared from the cache.
}
\description{
Clears the internal cache used by getGFishercoef_cached_cpp.
This can be useful to free memory or force recomputation.
}
