% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_coefficients_one_cpp}
\alias{compute_coefficients_one_cpp}
\title{Compute Coefficient Integrals for One-Sided P-Values (C++)}
\usage{
compute_coefficients_one_cpp(D, coeff_type)
}
\arguments{
\item{D}{Vector of degrees of freedom}

\item{coeff_type}{Integer: 1, 2, 3, or 4 for the Hermite polynomial order}
}
\value{
Vector of coefficient integrals
}
\description{
Fast C++ implementation for one-sided p-values.
Computes integrals of the form:
\deqn{I_j(k) = \int_{-8}^{8} F^{-1}_{d_j}(\Phi(x)) \phi(x) H_k(x) dx}
}
\keyword{internal}
