% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_coefficients_two_cpp}
\alias{compute_coefficients_two_cpp}
\title{Compute Coefficient Integrals for Two-Sided P-Values (C++)}
\usage{
compute_coefficients_two_cpp(D, coeff_type)
}
\arguments{
\item{D}{Vector of degrees of freedom}

\item{coeff_type}{Integer: 2, 4, 6, or 8 for the Hermite polynomial order}
}
\value{
Vector of coefficient integrals
}
\description{
Fast C++ implementation of coefficient calculations using Gauss-Legendre quadrature.
For two-sided p-values, computes integrals of the form:
\deqn{I_j(k) = \int_{-8}^{8} F^{-1}_{d_j}(\Phi(x^2)) \phi(x) H_k(x) dx}
}
\keyword{internal}
