% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_GFisher_cache_size}
\alias{get_GFisher_cache_size}
\title{Get GFisher Coefficient Cache Size}
\usage{
get_GFisher_cache_size()
}
\value{
The number of cached coefficient values.
}
\description{
Returns the current number of entries in the coefficient cache.
}
