% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_GFisher_cache_stats}
\alias{get_GFisher_cache_stats}
\title{Get GFisher Cache Statistics}
\usage{
get_GFisher_cache_stats()
}
\value{
A list containing cache statistics:
\itemize{
  \item size: Current number of entries
  \item max_size: Maximum cache size before automatic clearing
  \item memory_kb: Approximate memory usage in KB
  \item fill_ratio: Percentage of max capacity used
}
}
\description{
Returns detailed statistics about the coefficient cache.
}
